requirements.txt
web3==5.31.1
python-dotenv
web3connectpy